fix test pipeline, add build-push image running when Dockerfile changes, move tests CmakeLists to root, add Catch2 using lib/Catch2, remove date lib, update deps in Dockerfile, use cmake binary instead of source compile, compile clang+llvm 16 from source, cleanup cpp_tests, add land-cover-code option(defaulting to clc18_kode)
