+++ .github/workflows/main.yaml
@@ -1,4 +1,3 @@
-
 name: CI
 
 on:
@@ -7,14 +6,73 @@ on:
   pull_request:
     branches: [ master ]
 
+env:
+  REGISTRY: ghcr.io
+  RASPUTIN_DATA_DIR: examples/example_data
+
 jobs:
-  build:
+  check_changes:
     runs-on: ubuntu-latest
-    container:
-      image: dnumgis/rasputin:latest
+    outputs:
+      run_job: ${{ steps.check_dockerfile.outputs.run_job }}
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+        with:
+          fetch-depth: 2
+
+      - name: Check Dockerfile
+        id: check_dockerfile
+        run: |
+          if git diff --name-only HEAD^ HEAD | grep -q "^Dockerfile$"; then
+            echo "run_job=true" >> "$GITHUB_OUTPUT"
+          else
+            echo "run_job=false" >> "$GITHUB_OUTPUT"
+          fi
+
+  build-and-push-image:
+    runs-on: ubuntu-latest
+    needs: check_changes
+    if: ${{ needs.check_changes.outputs.run_job == 'true' }}
+    permissions:
+      contents: read
+      packages: write
 
     steps:
-    - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
+
+      - name: Login
+        uses: docker/login-action@v2
+        with:
+          registry: ${{ env.REGISTRY }}
+          username: ${{ github.actor }}
+          password: ${{ secrets.GITHUB_TOKEN }}
+
+      - name: Extract metadata
+        id: meta
+        uses: docker/metadata-action@v4
+        with:
+          images: ${{ env.REGISTRY }}/${{ github.repository }}_${{ github.ref_name }}
+          tags: |
+              type=raw,value=latest,enable=true
+
+      - name: Build and Push
+        uses: docker/build-push-action@v4
+        with:
+          context: .
+          push: true
+          tags: ${{ steps.meta.outputs.tags }}
+          labels: ${{ steps.meta.outputs.labels }}
+
+  build:
+    runs-on: ubuntu-latest
+    needs: [check_changes, build-and-push-image]
+    if: ${{ !failure() && !cancelled() }}
+    container:
+      image: ghcr.io/${{ github.repository }}_${{ github.ref_name }}
+
+    steps:
+    - uses: actions/checkout@v3
 
     - name: Copy dependencies
       run: |
@@ -22,11 +80,7 @@ jobs:
 
     - name: Install
       run: |
-          python3 setup.py install
-
-    - name: Set Data Dir
-      run: |
-          echo "::set-env name=RASPUTIN_DATA_DIR::/__w/rasputin/rasputin/examples/example_data"
+          pip -v install -e .
 
     - name: Rasputin Store
       run: |
@@ -34,11 +88,9 @@ jobs:
 
     - name: Rasputin Web
       run: |
-          rasputin_web -output examples/example_data/web/ingoya_test -material examples/example_data/materials/material.yaml  ingoya_test
+          rasputin_web -output examples/example_data/web/ingoya_test -material examples/example_data/materials/material.yaml ingoya_test
 
     - name: Run tests
       run: |
-          pytest tests/test_mesh.py
-          pytest tests/test_polygons.py
-          pytest tests/test_read_raster_file.py
-          pytest tests/test_tin_repository.py
+          pytest tests
+          ./build/test_sun_position
