+++ .gitignore
@@ -2,15 +2,15 @@ cmake-*
 .idea
 .env
 *.swp
-lib/CGAL*
-lib/pybind11
-lib/armadillo*
-src/rasputin/*so
-src/rasputin.egg-info
+lib/**/*
+**/*.egg-info
+**/*.so
 __pycache__
 build
 dist
 .DS_Store
-lib/boo*
-lib/geometry
-lib/build_*
+.envrc
+compile_commands.json
+.cache/
+.python-version
+*.eggs
