--- Pipfile
@@ -1,18 +1,0 @@
-[[source]]
-url = 'https://pypi.python.org/simple'
-verify_ssl = true
-name = 'pypi'
-
-[packages]
-numpy = '*'
-pyproj = '*'
-Pillow = '*'
-h5py = '*'
-lxml = '*'
-shapely = '*'
-descartes = '*'
-meshio = '*'
-rasputin = {path = "."}
-
-[dev-packages]
-pytest = '*'
