+++ setup.cfg
@@ -1,0 +1,56 @@
+[metadata]
+name=rasputin
+version=0.1
+author=Ola Skavhaug
+author_email=ola@xal.no
+description=A simple tool for constructing TINs from DEMs
+readme=README.md
+
+[options]
+zip_safe=False
+include_package_data=True
+package_dir=
+    =src
+packages=find:
+install_requires =
+    cython
+    numpy
+    pyproj
+    Pillow
+    h5py
+    lxml
+    shapely
+    descartes
+    meshio
+    pybind11
+setup_requires =
+    cython
+    h5py
+    pybind11
+python_requires=>=3.10
+
+[options.packages.find]
+where=src
+
+[options.extras_require]
+all=pytest
+ipy=ipython
+
+[options.package_data]
+web =
+    index.js
+    index.html
+    data.js
+    favicon.ico
+web.js =
+    three.min.js
+    dat.min.js
+web.js.controls =
+    OrbitControls.js
+    PointerLockControls.js
+
+[options.entry_points]
+console_scripts =
+    rasputin_web=rasputin.web_visualize:visualize_tin
+    rasputin_store=rasputin.application:store_tin
+    rasputin_shade=rasputin.application:compute_shades
