+++ setup.py
@@ -1,15 +1,13 @@
 import os
 import re
 import sys
-import sysconfig
 import platform
 import subprocess
+from pathlib import Path
 
 from distutils.version import LooseVersion
-from setuptools import setup, find_packages, Extension
+from setuptools import setup, Extension
 from setuptools.command.build_ext import build_ext
-
-web_templates = "src/rasputin/web/*"
 
 
 class CMakeExtension(Extension):
@@ -28,10 +26,13 @@ class CMakeBuild(build_ext):
                 ", ".join(e.name for e in self.extensions))
 
         if platform.system() == "Windows":
-            cmake_version = LooseVersion(re.search(r'version\s*([\d.]+)',
-                                         out.decode()).group(1))
-            if cmake_version < '3.1.0':
-                raise RuntimeError("CMake >= 3.1.0 is required on Windows")
+            cmake_version = re.search(r'version\s*([\d.]+)', out.decode())
+            if cmake_version is not None:
+                cmake_version = LooseVersion(cmake_version.group(1))
+                if cmake_version < '3.1.0':
+                    raise RuntimeError("CMake >= 3.1.0 is required on Windows")
+            else:
+                raise ValueError("cmake version not found")
 
         for ext in self.extensions:
             self.build_extension(ext)
@@ -39,8 +40,11 @@ class CMakeBuild(build_ext):
     def build_extension(self, ext):
         extdir = os.path.abspath(
             os.path.dirname(self.get_ext_fullpath(ext.name)))
-        cmake_args = ['-DCMAKE_LIBRARY_OUTPUT_DIRECTORY=' + extdir,
-                      '-DPYTHON_EXECUTABLE=' + sys.executable]
+        cmake_args = [
+            '-DCMAKE_LIBRARY_OUTPUT_DIRECTORY=' + extdir,
+            '-DPYTHON_EXECUTABLE=' + sys.executable,
+            '-DCMAKE_EXPORT_COMPILE_COMMANDS=ON'
+        ]
 
         cfg = 'Debug' if self.debug else 'Release'
         build_args = ['--config', cfg]
@@ -53,55 +57,35 @@ class CMakeBuild(build_ext):
                 cmake_args += ['-A', 'x64']
             build_args += ['--', '/m']
         else:
-            cmake_args += ['-DCMAKE_BUILD_TYPE=' + cfg]
-            build_args += ['--', '-j2']
+            cmake_args += ['-DCMAKE_BUILD_TYPE=' + cfg, f"-DCMAKE_PREFIX_PATH={Path.home() / '.local'};{Path().cwd() / 'lib'}"]
+            build_args += ['--', '-j4']
 
         env = os.environ.copy()
         env['CXXFLAGS'] = '{} -DVERSION_INFO=\\"{}\\"'.format(
             env.get('CXXFLAGS', ''),
-            self.distribution.get_version())
+            self.distribution.get_version()
+        )
+
         if not os.path.exists(self.build_temp):
             os.makedirs(self.build_temp)
-        subprocess.check_call(['cmake', ext.sourcedir] + cmake_args,
-                              cwd=self.build_temp, env=env)
-        subprocess.check_call(['cmake', '--build', '.'] + build_args,
-                              cwd=self.build_temp)
+
+        cmake_build_command = ['cmake', "-S", ".", "-B", "build"] + cmake_args
+        _compile_and_check(cmake_build_command, env=env)
+        _compile_and_check(["cmake", "--build", "build"] + build_args)
+
         print()  # Add an empty line for cleaner output
 
+def _compile_and_check(cmd, *args, **kwargs):
+        proc = subprocess.run(cmd, *args, **kwargs, check=True)
+        try:
+            proc.check_returncode()
+        except subprocess.CalledProcessError as err:
+            raise RuntimeError(f"{cmd} exited with {proc.returncode}\n{proc.stderr}") from err
+
 setup(
-    name='rasputin',
-    version='0.1',
-    author='Ola Skavhaug',
-    author_email='ola@xal.no',
-    description='A simple tool for constructing TINs from DEMs',
-    long_description='',
-    # tell setuptools to look for any packages under 'src'
-    packages=find_packages('src'),
-    install_requires=[
-          'numpy',
-          'pyproj',
-          'Pillow',
-          'h5py',
-          'lxml',
-          'shapely',
-          'descartes',
-          'meshio'
-    ],
-    # tell setuptools that all packages will be under the 'src' directory
-    # and nowhere else
-    package_dir={'':'src'},
-    data_files=[("rasputin/web", ["web/index.js", "web/index.html", "web/data.js", "web/favicon.ico"]),
-                ("rasputin/web/js", ["web/js/three.min.js", "web/js/dat.min.js"]),
-                ("rasputin/web/js/controls", ["web/js/controls/OrbitControls.js", "web/js/controls/PointerLockControls.js"])],
-    # add an extension module named 'python_cpp_example' to the package 
+    # add an extension module named 'python_cpp_example' to the package
     # 'python_cpp_example'
     ext_modules=[CMakeExtension('rasputin/rasputin')],
     # add custom build_ext command
     cmdclass=dict(build_ext=CMakeBuild),
-    zip_safe=False,
-    entry_points={
-        'console_scripts': ['rasputin_web = rasputin.web_visualize:visualize_tin',
-                            'rasputin_store = rasputin.application:store_tin',
-                            'rasputin_shade = rasputin.application:compute_shades']}
 )
-
