+++ src/rasputin/cpp/solar_position.h
@@ -1,20 +1,20 @@
 //
 // Created by Ola Skavhaug on 26/08/2019.
 //
-
 #pragma once
 
 #include <cmath>
 #include <chrono>
-#ifndef __clang__
-#include <date/date.h>
-#endif
 #include <ctime>
+#include <tuple>
+#include <vector>
 
 namespace rasputin::solar_position::collectors {
 
-auto azimuth_and_elevation() {
-    return [] (double e0, double Gamma, double Phi, const double alpha_mark, const double delta_mark, const double H_mark) {
+inline auto azimuth_and_elevation() {
+    return [] (
+        double e0, double Gamma, double Phi, const double alpha_mark, const double delta_mark, const double H_mark
+    ) {
         return std::make_tuple(Phi, e0);
     };
 }
@@ -24,16 +24,18 @@ auto azimuth_and_elevation() {
 namespace rasputin::solar_position::delta_t_calculator {
 const std::vector<double> _dts{31.1, 33.2, 35.7, 40.2, 45.5, 50.5, 54.3, 56.9, 60.8, 63.8, 64.7};
 
-auto parabolic_dt_calc() {
+inline auto parabolic_dt_calc() {
     return [] (const double year) {
          return -20 + 32*pow((year - 1820)/20.0, 2);
     };
 }
 
-auto coarse_date_calc() {
-    return [] (const unsigned int year,
-               const unsigned int month,
-               const unsigned int day) {
+inline auto coarse_date_calc() {
+    return [] (
+        const unsigned int year,
+        const unsigned int month,
+        const unsigned int day
+    ) {
         if (year < 1955 || year > 2009)
             return parabolic_dt_calc()(year);
         const auto idx = (year - 1955) / 5;
@@ -41,13 +43,10 @@ auto coarse_date_calc() {
     };
 }
 
-auto coarse_timestamp_calc() {
+inline auto coarse_timestamp_calc() {
     using namespace std::chrono;
-#ifndef __clang__
-    using namespace date;
-#endif
 
-    auto ep = sys_days{January/1/1970};
+    auto ep = sys_days{1970y / January / 1};
 
     return [ep] (system_clock::time_point tp) {
         const auto year = round(duration_cast<seconds>(tp - ep).count()*1.0/duration_cast<seconds>(years(1)).count()) + 1970;
@@ -62,19 +61,19 @@ auto coarse_timestamp_calc() {
 
 namespace rasputin::solar_position {
 
-auto limit_degrees(const double degrees) {
+inline double limit_degrees(const double degrees) {
     return fmod(fmod(degrees, 360) + 360, 360);
 }
 
-auto d2r(const double d) {
+inline double d2r(const double d) {
     return d*M_PI/180.0;
 }
 
-auto r2d(const double r) {
+inline double r2d(const double r) {
     return r*180.0/M_PI;
 }
 
-auto jd_from_cal(unsigned int year, unsigned int month, const double day){
+inline double jd_from_cal(unsigned int year, unsigned int month, const double day){
 
     if (month < 3) {
         month += 12;
@@ -82,40 +81,39 @@ auto jd_from_cal(unsigned int year, unsigned int month, const double day){
     }
 
     // Equation (4)
-    const auto jd_temp = trunc(365.25*(year + 4716)) + trunc(30.6001*(month + 1)) + day - 1524.5;
+    const double jd_temp = trunc(365.25*(year + 4716)) + trunc(30.6001*(month + 1)) + day - 1524.5;
     if (jd_temp < 2299160)
         return jd_temp;
-    const auto A = trunc(year/100);
+    const double A = trunc(year/100.);
     return jd_temp + 2 - A + trunc(A/4);
 }
 
-double jd_from_clock(const std::chrono::system_clock::time_point tp) {
+inline double jd_from_clock(const std::chrono::system_clock::time_point tp) {
     using namespace std::chrono;
-#ifndef __clang__
-    using namespace date;
-#endif
-    const auto dt = tp - sys_days{January/1/1970};
-    const auto t = duration_cast<seconds>(dt).count();
+
+    const auto dt = tp - sys_days{1970y / January / 1};
+    const long t = duration_cast<seconds>(dt).count();
+
     return t/86400.0 + 2440587.5;
 }
 
-double jde(const double julian_day, const double dt){
+inline double jde(const double julian_day, const double dt){
     return julian_day + dt/86400.0;
 }
 
-double jc(const double julian_day) {
+inline double jc(const double julian_day) {
     return (julian_day - 2451545.0)/36525.0;
 }
 
-double jce(const double julian_ephemeris_day) {
+inline double jce(const double julian_ephemeris_day) {
     return (julian_ephemeris_day - 2451545.0)/36525.0;
 }
 
-double jme(const double julian_ephemeris_centry) {
+inline double jme(const double julian_ephemeris_centry) {
     return julian_ephemeris_centry/10.0;
 }
 
-double heliocentric_longitude(const double julian_ephemeris_millennium) {
+inline double heliocentric_longitude(const double julian_ephemeris_millennium) {
     const double jme = julian_ephemeris_millennium;
     const double L00  = 175347046.0;
     const double L01  = 3341656*cos(4.6692568 +  6283.07585 *jme);
@@ -280,7 +278,7 @@ double heliocentric_longitude(const double julian_ephemeris_millennium) {
     return limit_degrees(L_deg);
 }
 
-double heliocentric_latitude(double julian_ephemeris_millennium) {
+inline double heliocentric_latitude(double julian_ephemeris_millennium) {
     const double jme = julian_ephemeris_millennium;
 
     const double B00 = 280*cos(3.199 + 84334.662*jme);
@@ -289,7 +287,7 @@ double heliocentric_latitude(double julian_ephemeris_millennium) {
     const double B03 =  44*cos(3.7   +  2352.87 *jme);
     const double B04 =  32*cos(4     +  1577.34 *jme);
     const double B0 = B00 + B01 + B02 + B03 + B04;
-    
+
     const double B10 = 9*cos(3.9  + 5507.55*jme);
     const double B11 = 6*cos(1.73 + 5223.69*jme);
     const double B1 = B10 + B11;
@@ -299,7 +297,7 @@ double heliocentric_latitude(double julian_ephemeris_millennium) {
     return B_deg;
 }
 
-double heliocentric_radius_vector(double julian_ephemeris_millennium) {
+inline double heliocentric_radius_vector(double julian_ephemeris_millennium) {
     const double jme = julian_ephemeris_millennium;
 
     const double R00  = 100013989;
@@ -375,130 +373,130 @@ double heliocentric_radius_vector(double julian_ephemeris_millennium) {
     return (R0 + R1*jme + R2*pow(jme, 2) + R3*pow(jme, 3) + R4*pow(jme, 4))*1e-8;
 }
 
-double geocentric_longitude(const double heliocentric_longitude) {
+inline double geocentric_longitude(const double heliocentric_longitude) {
     return heliocentric_longitude + 180;
 }
 
-double geocentric_latitude(const double heliocentric_latitude) {
+inline double geocentric_latitude(const double heliocentric_latitude) {
     return -heliocentric_latitude;
 }
 
 
-auto nutation(const double julian_ephemeris_centry) {
-    const auto jce = julian_ephemeris_centry;
-    const auto jce2 = jce*jce;
-    const auto jce3 = jce2*jce;
+inline std::pair<double, double> nutation(const double julian_ephemeris_centry) {
+    const double jce = julian_ephemeris_centry;
+    const double jce2 = jce*jce;
+    const double jce3 = jce2*jce;
 
-    const auto X0 = 297.85036 + 445267.111480*jce - 0.0019142*jce2 + jce3/189474;
-    const auto X1 = 357.52772 +  35999.050340*jce - 0.0001603*jce2 - jce3/300000;
-    const auto X2 = 134.96298 + 477198.867398*jce + 0.0086972*jce2 + jce3/56250;
-    const auto X3 = 93.27191  + 483202.017538*jce - 0.0036825*jce2 + jce3/327270;
-    const auto X4 = 125.04452 -   1934.136261*jce + 0.0020708*jce2 + jce3/450000;
+    const double X0 = 297.85036 + 445267.111480*jce - 0.0019142*jce2 + jce3/189474;
+    const double X1 = 357.52772 +  35999.050340*jce - 0.0001603*jce2 - jce3/300000;
+    const double X2 = 134.96298 + 477198.867398*jce + 0.0086972*jce2 + jce3/56250;
+    const double X3 = 93.27191  + 483202.017538*jce - 0.0036825*jce2 + jce3/327270;
+    const double X4 = 125.04452 -   1934.136261*jce + 0.0020708*jce2 + jce3/450000;
 
-    const auto dksi_00 = (-171996 - 174.2*jce)*sin(d2r(                        X4  ));
-    const auto dksi_01 = ( -13187 -   1.6*jce)*sin(d2r(-X0*2            +X3*2 +X4*2));
-    const auto dksi_02 = (  -2274 -   0.2*jce)*sin(d2r(                  X3*2 +X4*2));
-    const auto dksi_03 = (   2062 +   0.2*jce)*sin(d2r(                        X4*2));
-    const auto dksi_04 = (   1426 -   3.4*jce)*sin(d2r(       X1                   ));
-    const auto dksi_05 = (    712 +   0.1*jce)*sin(d2r(            X2              ));
-    const auto dksi_06 = (   -517 +   1.2*jce)*sin(d2r(-X0*2 +X1        +X3*2 +X4*2));
-    const auto dksi_07 = (   -386 -   0.4*jce)*sin(d2r(                  X3*2 +X4  ));
-    const auto dksi_08 =     -301             *sin(d2r(            X2   +X3*2 +X4*2));
-    const auto dksi_09 = (    217 -   0.5*jce)*sin(d2r(-X0*2 -X1        +X3*2 +X4*2));
-    const auto dksi_10 =     -158             *sin(d2r(-X0*2      +X2              ));
-    const auto dksi_11 = (    129 +   0.1*jce)*sin(d2r(-X0*2            +X3*2 +X4  ));
-    const auto dksi_12 =      123             *sin(d2r(           -X2   +X3*2 +X4*2));
-    const auto dksi_13 =       63             *sin(d2r( X0*2                       ));
-    const auto dksi_14 = (     63 +   0.1*jce)*sin(d2r(            X2         +X4  ));
-    const auto dksi_15 =      -59             *sin(d2r( X0*2      -X2   +X3*2 +X4*2));
-    const auto dksi_16 = (    -58 -   0.1*jce)*sin(d2r(           -X2         +X4  ));
-    const auto dksi_17 =      -51             *sin(d2r(            X2   +X3*2 +X4  ));
-    const auto dksi_18 =       48             *sin(d2r(-X0*2      +X2*2            ));
-    const auto dksi_19 =       46             *sin(d2r(           -X2*2 +X3*2 +X4  ));
-    const auto dksi_20 =      -38             *sin(d2r( X0*2            +X3*2 +X4*2));
-    const auto dksi_21 =      -31             *sin(d2r(            X2*2 +X3*2 +X4*2));
-    const auto dksi_22 =       29             *sin(d2r(            X2*2            ));
-    const auto dksi_23 =       29             *sin(d2r(-X0*2      +X2   +X3*2 +X4*2));
-    const auto dksi_24 =       26             *sin(d2r(                  X3*2      ));
-    const auto dksi_25 =      -22             *sin(d2r(-X0*2            +X3*2      ));
-    const auto dksi_26 =       21             *sin(d2r(           -X2   +X3*2 +X4  ));
-    const auto dksi_27 = (     17 -   0.1*jce)*sin(d2r(       X1*2                 ));
-    const auto dksi_28 =       16             *sin(d2r( X0*2      -X2         +X4  ));
-    const auto dksi_29 = (    -16 +   0.1*jce)*sin(d2r(-X0*2 +X1*2      +X3*2 +X4*2));
-    const auto dksi_30 =      -15             *sin(d2r(       X1              +X4  ));
-    const auto dksi_31 =      -13             *sin(d2r(-X0*2      +X2         +X4  ));
-    const auto dksi_32 =      -12             *sin(d2r(      -X1              +X4  ));
-    const auto dksi_33 =       11             *sin(d2r(            X2*2 -X3*2      ));
-    const auto dksi_34 =      -10             *sin(d2r( X0*2      -X2   +X3*2 +X4  ));
-    const auto dksi_35 =       -8             *sin(d2r( X0*2      +X2   +X3*2 +X4*2));
-    const auto dksi_36 =        7             *sin(d2r(       X1        +X3*2 +X4*2));
-    const auto dksi_37 =       -7             *sin(d2r(-X0*2 +X1  +X2              ));
-    const auto dksi_38 =       -7             *sin(d2r(      -X1        +X3*2 +X4*2));
-    const auto dksi_39 =       -7             *sin(d2r( X0*2            +X3*2 +X4  ));
-    const auto dksi_40 =        6             *sin(d2r( X0*2      +X2              ));
-    const auto dksi_41 =        6             *sin(d2r(-X0*2      +X2*2 +X3*2 +X4*2));
-    const auto dksi_42 =        6             *sin(d2r(-X0*2      +X2   +X3*2 +X4  ));
-    const auto dksi_43 =       -6             *sin(d2r( X0*2      -X2*2       +X4  ));
-    const auto dksi_44 =       -6             *sin(d2r( X0*2                  +X4  ));
-    const auto dksi_45 =        5             *sin(d2r(      -X1  +X2              ));
-    const auto dksi_46 =       -5             *sin(d2r(-X0*2 -X1        +X3*2 +X4  ));
-    const auto dksi_47 =       -5             *sin(d2r(-X0*2                  +X4  ));
-    const auto dksi_48 =       -5             *sin(d2r(            X2*2 +X3*2 +X4  ));
-    const auto dksi_49 =        4             *sin(d2r(-X0*2      +X2*2       +X4  ));
-    const auto dksi_50 =        4             *sin(d2r(-X0*2 +X1        +X3*2 +X4  ));
-    const auto dksi_51 =        4             *sin(d2r(            X2   -X3*2      ));
-    const auto dksi_52 =       -4             *sin(d2r(-X0        +X2              ));
-    const auto dksi_53 =       -4             *sin(d2r(-X0*2 +X1                   ));
-    const auto dksi_54 =       -4             *sin(d2r( X0                         ));
-    const auto dksi_55 =        3             *sin(d2r(            X2   +X3*2      ));
-    const auto dksi_56 =       -3             *sin(d2r(           -X2*2 +X3*2 +X4*2));
-    const auto dksi_57 =       -3             *sin(d2r(-X0   -X1  +X2              ));
-    const auto dksi_58 =       -3             *sin(d2r(       X1  +X2              ));
-    const auto dksi_59 =       -3             *sin(d2r(      -X1  +X2   +X3*2 +X4*2));
-    const auto dksi_60 =       -3             *sin(d2r( X0*2 -X1  -X2   +X3*2 +X4*2));
-    const auto dksi_61 =       -3             *sin(d2r(            X2*3 +X3*2 +X4*2));
-    const auto dksi_62 =       -3             *sin(d2r( X0*2 -X1        +X3*2 +X4*2));
+    const double dksi_00 = (-171996 - 174.2*jce)*sin(d2r(                        X4  ));
+    const double dksi_01 = ( -13187 -   1.6*jce)*sin(d2r(-X0*2            +X3*2 +X4*2));
+    const double dksi_02 = (  -2274 -   0.2*jce)*sin(d2r(                  X3*2 +X4*2));
+    const double dksi_03 = (   2062 +   0.2*jce)*sin(d2r(                        X4*2));
+    const double dksi_04 = (   1426 -   3.4*jce)*sin(d2r(       X1                   ));
+    const double dksi_05 = (    712 +   0.1*jce)*sin(d2r(            X2              ));
+    const double dksi_06 = (   -517 +   1.2*jce)*sin(d2r(-X0*2 +X1        +X3*2 +X4*2));
+    const double dksi_07 = (   -386 -   0.4*jce)*sin(d2r(                  X3*2 +X4  ));
+    const double dksi_08 =     -301             *sin(d2r(            X2   +X3*2 +X4*2));
+    const double dksi_09 = (    217 -   0.5*jce)*sin(d2r(-X0*2 -X1        +X3*2 +X4*2));
+    const double dksi_10 =     -158             *sin(d2r(-X0*2      +X2              ));
+    const double dksi_11 = (    129 +   0.1*jce)*sin(d2r(-X0*2            +X3*2 +X4  ));
+    const double dksi_12 =      123             *sin(d2r(           -X2   +X3*2 +X4*2));
+    const double dksi_13 =       63             *sin(d2r( X0*2                       ));
+    const double dksi_14 = (     63 +   0.1*jce)*sin(d2r(            X2         +X4  ));
+    const double dksi_15 =      -59             *sin(d2r( X0*2      -X2   +X3*2 +X4*2));
+    const double dksi_16 = (    -58 -   0.1*jce)*sin(d2r(           -X2         +X4  ));
+    const double dksi_17 =      -51             *sin(d2r(            X2   +X3*2 +X4  ));
+    const double dksi_18 =       48             *sin(d2r(-X0*2      +X2*2            ));
+    const double dksi_19 =       46             *sin(d2r(           -X2*2 +X3*2 +X4  ));
+    const double dksi_20 =      -38             *sin(d2r( X0*2            +X3*2 +X4*2));
+    const double dksi_21 =      -31             *sin(d2r(            X2*2 +X3*2 +X4*2));
+    const double dksi_22 =       29             *sin(d2r(            X2*2            ));
+    const double dksi_23 =       29             *sin(d2r(-X0*2      +X2   +X3*2 +X4*2));
+    const double dksi_24 =       26             *sin(d2r(                  X3*2      ));
+    const double dksi_25 =      -22             *sin(d2r(-X0*2            +X3*2      ));
+    const double dksi_26 =       21             *sin(d2r(           -X2   +X3*2 +X4  ));
+    const double dksi_27 = (     17 -   0.1*jce)*sin(d2r(       X1*2                 ));
+    const double dksi_28 =       16             *sin(d2r( X0*2      -X2         +X4  ));
+    const double dksi_29 = (    -16 +   0.1*jce)*sin(d2r(-X0*2 +X1*2      +X3*2 +X4*2));
+    const double dksi_30 =      -15             *sin(d2r(       X1              +X4  ));
+    const double dksi_31 =      -13             *sin(d2r(-X0*2      +X2         +X4  ));
+    const double dksi_32 =      -12             *sin(d2r(      -X1              +X4  ));
+    const double dksi_33 =       11             *sin(d2r(            X2*2 -X3*2      ));
+    const double dksi_34 =      -10             *sin(d2r( X0*2      -X2   +X3*2 +X4  ));
+    const double dksi_35 =       -8             *sin(d2r( X0*2      +X2   +X3*2 +X4*2));
+    const double dksi_36 =        7             *sin(d2r(       X1        +X3*2 +X4*2));
+    const double dksi_37 =       -7             *sin(d2r(-X0*2 +X1  +X2              ));
+    const double dksi_38 =       -7             *sin(d2r(      -X1        +X3*2 +X4*2));
+    const double dksi_39 =       -7             *sin(d2r( X0*2            +X3*2 +X4  ));
+    const double dksi_40 =        6             *sin(d2r( X0*2      +X2              ));
+    const double dksi_41 =        6             *sin(d2r(-X0*2      +X2*2 +X3*2 +X4*2));
+    const double dksi_42 =        6             *sin(d2r(-X0*2      +X2   +X3*2 +X4  ));
+    const double dksi_43 =       -6             *sin(d2r( X0*2      -X2*2       +X4  ));
+    const double dksi_44 =       -6             *sin(d2r( X0*2                  +X4  ));
+    const double dksi_45 =        5             *sin(d2r(      -X1  +X2              ));
+    const double dksi_46 =       -5             *sin(d2r(-X0*2 -X1        +X3*2 +X4  ));
+    const double dksi_47 =       -5             *sin(d2r(-X0*2                  +X4  ));
+    const double dksi_48 =       -5             *sin(d2r(            X2*2 +X3*2 +X4  ));
+    const double dksi_49 =        4             *sin(d2r(-X0*2      +X2*2       +X4  ));
+    const double dksi_50 =        4             *sin(d2r(-X0*2 +X1        +X3*2 +X4  ));
+    const double dksi_51 =        4             *sin(d2r(            X2   -X3*2      ));
+    const double dksi_52 =       -4             *sin(d2r(-X0        +X2              ));
+    const double dksi_53 =       -4             *sin(d2r(-X0*2 +X1                   ));
+    const double dksi_54 =       -4             *sin(d2r( X0                         ));
+    const double dksi_55 =        3             *sin(d2r(            X2   +X3*2      ));
+    const double dksi_56 =       -3             *sin(d2r(           -X2*2 +X3*2 +X4*2));
+    const double dksi_57 =       -3             *sin(d2r(-X0   -X1  +X2              ));
+    const double dksi_58 =       -3             *sin(d2r(       X1  +X2              ));
+    const double dksi_59 =       -3             *sin(d2r(      -X1  +X2   +X3*2 +X4*2));
+    const double dksi_60 =       -3             *sin(d2r( X0*2 -X1  -X2   +X3*2 +X4*2));
+    const double dksi_61 =       -3             *sin(d2r(            X2*3 +X3*2 +X4*2));
+    const double dksi_62 =       -3             *sin(d2r( X0*2 -X1        +X3*2 +X4*2));
 
-    const auto deps_00 = (92025 + 8.9*jce)*cos(d2r(                          X4  ));
-    const auto deps_01 = ( 5736 - 3.1*jce)*cos(d2r(-X0*2              +X3*2 +X4*2));
-    const auto deps_02 = (  977 - 0.5*jce)*cos(d2r(                    X3*2 +X4*2));
-    const auto deps_03 = ( -895 + 0.5*jce)*cos(d2r(                          X4*2));
-    const auto deps_04 = (   54 - 0.1*jce)*cos(d2r(       X1                     ));
-    const auto deps_05 =     -7           *cos(d2r(              X2              ));
-    const auto deps_06 = (  224 - 0.6*jce)*cos(d2r(-X0*2 +X1          +X3*2 +X4*2));
-    const auto deps_07 =    200           *cos(d2r(                    X3*2 +X4  ));
-    const auto deps_08 = (  129 - 0.1*jce)*cos(d2r(              X2   +X3*2 +X4*2));
-    const auto deps_09 = (  -95 + 0.3*jce)*cos(d2r(-X0*2 -X1          +X3*2 +X4*2));
-    const auto deps_10 =    -70           *cos(d2r(-X0*2              +X3*2 +X4  ));
-    const auto deps_11 =    -53           *cos(d2r(             -X2   +X3*2 +X4*2));
-    const auto deps_12 =    -33           *cos(d2r(              X2         +X4  ));
-    const auto deps_13 =     26           *cos(d2r( X0*2        -X2   +X3*2 +X4*2));
-    const auto deps_14 =     32           *cos(d2r(             -X2         +X4  ));
-    const auto deps_15 =     27           *cos(d2r(              X2   +X3*2 +X4  ));
-    const auto deps_16 =    -24           *cos(d2r(             -X2*2 +X3*2 +X4  ));
-    const auto deps_17 =     16           *cos(d2r( X0*2              +X3*2 +X4*2));
-    const auto deps_18 =     13           *cos(d2r(              X2*2 +X3*2 +X4*2));
-    const auto deps_19 =    -12           *cos(d2r(-X0*2        +X2   +X3*2 +X4*2));
-    const auto deps_20 =    -10           *cos(d2r(             -X2   +X3*2 +X4  ));
-    const auto deps_21 =     -8           *cos(d2r( X0*2        -X2         +X4  ));
-    const auto deps_22 =      7           *cos(d2r(-X0*2 +X1*2        +X3*2 +X4*2));
-    const auto deps_23 =      9           *cos(d2r(       X1                +X4  ));
-    const auto deps_24 =      7           *cos(d2r(-X0*2        +X2         +X4  ));
-    const auto deps_25 =      6           *cos(d2r(      -X1                +X4  ));
-    const auto deps_26 =      5           *cos(d2r( X0*2        -X2   +X3*2 +X4  ));
-    const auto deps_27 =      3           *cos(d2r( X0*2        +X2   +X3*2 +X4*2));
-    const auto deps_28 =     -3           *cos(d2r(       X1          +X3*2 +X4*2));
-    const auto deps_29 =      3           *cos(d2r(      -X1          +X3*2 +X4*2));
-    const auto deps_30 =      3           *cos(d2r( X0*2              +X3*2 +X4  ));
-    const auto deps_31 =     -3           *cos(d2r(-X0*2        +X2*2 +X3*2 +X4*2));
-    const auto deps_32 =     -3           *cos(d2r(-X0*2        +X2   +X3*2 +X4  ));
-    const auto deps_33 =      3           *cos(d2r( X0*2        -X2*2       +X4  ));
-    const auto deps_34 =      3           *cos(d2r( X0*2                    +X4  ));
-    const auto deps_35 =      3           *cos(d2r(-X0*2 -X1          +X3*2 +X4  ));
-    const auto deps_36 =      3           *cos(d2r(-X0*2                    +X4  ));
-    const auto deps_37 =      3           *cos(d2r(              X2*2 +X3*2 +X4  ));
+    const double deps_00 = (92025 + 8.9*jce)*cos(d2r(                          X4  ));
+    const double deps_01 = ( 5736 - 3.1*jce)*cos(d2r(-X0*2              +X3*2 +X4*2));
+    const double deps_02 = (  977 - 0.5*jce)*cos(d2r(                    X3*2 +X4*2));
+    const double deps_03 = ( -895 + 0.5*jce)*cos(d2r(                          X4*2));
+    const double deps_04 = (   54 - 0.1*jce)*cos(d2r(       X1                     ));
+    const double deps_05 =     -7           *cos(d2r(              X2              ));
+    const double deps_06 = (  224 - 0.6*jce)*cos(d2r(-X0*2 +X1          +X3*2 +X4*2));
+    const double deps_07 =    200           *cos(d2r(                    X3*2 +X4  ));
+    const double deps_08 = (  129 - 0.1*jce)*cos(d2r(              X2   +X3*2 +X4*2));
+    const double deps_09 = (  -95 + 0.3*jce)*cos(d2r(-X0*2 -X1          +X3*2 +X4*2));
+    const double deps_10 =    -70           *cos(d2r(-X0*2              +X3*2 +X4  ));
+    const double deps_11 =    -53           *cos(d2r(             -X2   +X3*2 +X4*2));
+    const double deps_12 =    -33           *cos(d2r(              X2         +X4  ));
+    const double deps_13 =     26           *cos(d2r( X0*2        -X2   +X3*2 +X4*2));
+    const double deps_14 =     32           *cos(d2r(             -X2         +X4  ));
+    const double deps_15 =     27           *cos(d2r(              X2   +X3*2 +X4  ));
+    const double deps_16 =    -24           *cos(d2r(             -X2*2 +X3*2 +X4  ));
+    const double deps_17 =     16           *cos(d2r( X0*2              +X3*2 +X4*2));
+    const double deps_18 =     13           *cos(d2r(              X2*2 +X3*2 +X4*2));
+    const double deps_19 =    -12           *cos(d2r(-X0*2        +X2   +X3*2 +X4*2));
+    const double deps_20 =    -10           *cos(d2r(             -X2   +X3*2 +X4  ));
+    const double deps_21 =     -8           *cos(d2r( X0*2        -X2         +X4  ));
+    const double deps_22 =      7           *cos(d2r(-X0*2 +X1*2        +X3*2 +X4*2));
+    const double deps_23 =      9           *cos(d2r(       X1                +X4  ));
+    const double deps_24 =      7           *cos(d2r(-X0*2        +X2         +X4  ));
+    const double deps_25 =      6           *cos(d2r(      -X1                +X4  ));
+    const double deps_26 =      5           *cos(d2r( X0*2        -X2   +X3*2 +X4  ));
+    const double deps_27 =      3           *cos(d2r( X0*2        +X2   +X3*2 +X4*2));
+    const double deps_28 =     -3           *cos(d2r(       X1          +X3*2 +X4*2));
+    const double deps_29 =      3           *cos(d2r(      -X1          +X3*2 +X4*2));
+    const double deps_30 =      3           *cos(d2r( X0*2              +X3*2 +X4  ));
+    const double deps_31 =     -3           *cos(d2r(-X0*2        +X2*2 +X3*2 +X4*2));
+    const double deps_32 =     -3           *cos(d2r(-X0*2        +X2   +X3*2 +X4  ));
+    const double deps_33 =      3           *cos(d2r( X0*2        -X2*2       +X4  ));
+    const double deps_34 =      3           *cos(d2r( X0*2                    +X4  ));
+    const double deps_35 =      3           *cos(d2r(-X0*2 -X1          +X3*2 +X4  ));
+    const double deps_36 =      3           *cos(d2r(-X0*2                    +X4  ));
+    const double deps_37 =      3           *cos(d2r(              X2*2 +X3*2 +X4  ));
 
-    const auto dksi = (dksi_00 + dksi_01 + dksi_02 + dksi_03 + dksi_04 + dksi_05 + dksi_06 + dksi_07 + dksi_08 + dksi_09
+    const double dksi = (dksi_00 + dksi_01 + dksi_02 + dksi_03 + dksi_04 + dksi_05 + dksi_06 + dksi_07 + dksi_08 + dksi_09
                      + dksi_10 + dksi_11 + dksi_12 + dksi_13 + dksi_14 + dksi_15 + dksi_16 + dksi_17 + dksi_18 + dksi_19
                      + dksi_20 + dksi_21 + dksi_22 + dksi_23 + dksi_24 + dksi_25 + dksi_26 + dksi_27 + dksi_28 + dksi_29
                      + dksi_30 + dksi_31 + dksi_32 + dksi_33 + dksi_34 + dksi_35 + dksi_36 + dksi_37 + dksi_38 + dksi_39
@@ -506,7 +504,7 @@ auto nutation(const double julian_ephemeris_centry) {
                      + dksi_50 + dksi_51 + dksi_52 + dksi_53 + dksi_54 + dksi_55 + dksi_56 + dksi_57 + dksi_58 + dksi_59
                      + dksi_60 + dksi_61 + dksi_62)/36000000.0;
 
-    const auto deps = (deps_00 + deps_01 + deps_02 + deps_03 + deps_04 + deps_05 + deps_06 + deps_07 + deps_08 + deps_09
+    const double deps = (deps_00 + deps_01 + deps_02 + deps_03 + deps_04 + deps_05 + deps_06 + deps_07 + deps_08 + deps_09
                      + deps_10 + deps_11 + deps_12 + deps_13 + deps_14 + deps_15 + deps_16 + deps_17 + deps_18 + deps_19
                      + deps_20 + deps_21 + deps_22 + deps_23 + deps_24 + deps_25 + deps_26 + deps_27 + deps_28 + deps_29
                      + deps_30 + deps_31 + deps_32 + deps_33 + deps_34 + deps_35 + deps_36 + deps_37)/36000000.0;
@@ -514,8 +512,7 @@ auto nutation(const double julian_ephemeris_centry) {
     return std::make_pair(dksi, deps);
 }
 
-double true_ecliptic_obliquity(const double julian_ephemeris_millennium,
-                               const double obliquity_nutation) {
+inline double true_ecliptic_obliquity(const double julian_ephemeris_millennium, const double obliquity_nutation) {
     const double U = julian_ephemeris_millennium/10;
     const double U2 = U*U;
     const double U3 = U2*U;
@@ -527,70 +524,81 @@ double true_ecliptic_obliquity(const double julian_ephemeris_millennium,
     const double U9 = U8*U;
     const double U10 = U9*U;
     // Equation (24)
-    const auto eps_0 = 84381.448 - 4680.93*U   - 1.55*U2 + 1999.25*U3
+    const double eps_0 = 84381.448 - 4680.93*U   - 1.55*U2 + 1999.25*U3
                         - 51.38*U4 -249.67*U5 - 39.05*U6 + 7.12*U7
                         + 27.87*U8 +  5.79*U9 + 2.45*U10;
     // Equation (25)
     return eps_0/3600.0 + obliquity_nutation;
 }
 
-double aberration_correction(const double R) {
+inline double aberration_correction(const double R) {
     // Equation (26)
     return -20.4898/(3600.0*R);
 }
 
-double apparent_sun_longitude(const double geoc_long, const double nut_long,  const double ab_corr) {
+inline double apparent_sun_longitude(const double geoc_long, const double nut_long, const double ab_corr) {
     // Equation (27)
     return geoc_long + nut_long + ab_corr;
 }
-double apparent_Greenwich_sidereal_time(const double julian_day,
-                                              const double julian_ephemeris_centry,
-                                              const double nut_long,
-                                              const double true_ecliptic_obliquity) {
+
+inline double apparent_Greenwich_sidereal_time(
+    const double julian_day,
+    const double julian_ephemeris_centry,
+    const double nut_long,
+    const double true_ecliptic_obliquity
+) {
     // Equation (28)
-    const double nu0 = 280.46061837 + 360.98564736629*(julian_day - 2451545) + 0.000387933*pow(julian_ephemeris_centry, 2) -
-                     pow(julian_ephemeris_centry, 3)/38710000;
+    const double nu0 =
+        280.46061837
+        + 360.98564736629*(julian_day - 2451545)
+        + 0.000387933*pow(julian_ephemeris_centry, 2)
+        - pow(julian_ephemeris_centry, 3)/38710000;
 
     // Equation (29) with [0, 360) range limit
     return limit_degrees(nu0) + nut_long*cos(d2r(true_ecliptic_obliquity));
 }
 
-double geocentric_sun_right_ascension(const double app_sun_longitude,
-                           const double true_ecliptic_obliq,
-                           const double geocentric_lat) {
+inline double geocentric_sun_right_ascension(
+    const double app_sun_longitude, const double true_ecliptic_obliq, const double geocentric_lat
+) {
     // Equation (30)
-    const auto lambda_rad = d2r(app_sun_longitude);
-    const auto eps_rad = d2r(true_ecliptic_obliq);
-    const auto beta_rad = d2r(geocentric_lat);
-    const auto alpha = atan2(sin(lambda_rad)*cos(eps_rad) - tan(beta_rad)*sin(eps_rad), cos(lambda_rad));
+    const double lambda_rad = d2r(app_sun_longitude);
+    const double eps_rad = d2r(true_ecliptic_obliq);
+    const double beta_rad = d2r(geocentric_lat);
+    const double alpha = atan2(sin(lambda_rad)*cos(eps_rad) - tan(beta_rad)*sin(eps_rad), cos(lambda_rad));
+
     return limit_degrees(r2d(alpha));
 }
 
-double geocentric_sun_declination(const double app_sun_longitude,
-                                  const double true_ecliptic_obliq,
-                                  const double geocentric_lat) {
+inline double geocentric_sun_declination(
+    const double app_sun_longitude, const double true_ecliptic_obliq, const double geocentric_lat
+) {
     // Equation (31)
-    const auto lambda_rad = d2r(app_sun_longitude);
-    const auto eps_rad = d2r(true_ecliptic_obliq);
-    const auto beta_rad = d2r(geocentric_lat);
-    const auto delta = asin(sin(beta_rad)*cos(eps_rad) + cos(beta_rad)*sin(eps_rad)*sin(lambda_rad));
+    const double lambda_rad = d2r(app_sun_longitude);
+    const double eps_rad = d2r(true_ecliptic_obliq);
+    const double beta_rad = d2r(geocentric_lat);
+    const double delta = asin(sin(beta_rad)*cos(eps_rad) + cos(beta_rad)*sin(eps_rad)*sin(lambda_rad));
+
     return r2d(delta);
 }
 
-double observer_local_hour_angle(const double app_Greenwich_sidereal_time,
-                        const double geocentric_sun_right_asc,
-                        const double geographical_lon) {
+inline double observer_local_hour_angle(
+    const double app_Greenwich_sidereal_time, const double geocentric_sun_right_asc, const double geographical_lon
+) {
     // Equation (32)
     const double H = app_Greenwich_sidereal_time + geographical_lon - geocentric_sun_right_asc;
+
     return limit_degrees(H);
 }
 
-auto topocentric_values(const double R,
-                        const double geographical_latitude,
-                        const double masl,
-                        const double geocentric_sun_right_asc,
-                        const double geocentric_sun_decl,
-                        const double obs_local_hour_angle) {
+inline std::tuple<double, double, double> topocentric_values(
+    const double R,
+    const double geographical_latitude,
+    const double masl,
+    const double geocentric_sun_right_asc,
+    const double geocentric_sun_decl,
+    const double obs_local_hour_angle
+) {
 
     // Convert to rad
     const double lat_rad = d2r(geographical_latitude);
@@ -622,138 +630,133 @@ auto topocentric_values(const double R,
     return std::make_tuple(topo_sun_right_asc , topo_sun_declination, topo_local_hour_angle);
 }
 
-auto uncorrected_topocentric_elevation_angle(const double geographical_latitude,
-                                             const double topo_sun_declination,
-                                             const double topo_local_hour_angle) {
+inline double uncorrected_topocentric_elevation_angle(
+    const double geographical_latitude, const double topo_sun_declination, const double topo_local_hour_angle
+) {
     // Equation (41)
     const double lat_rad = d2r(geographical_latitude);
     const double sun_decl_rad = d2r(topo_sun_declination);
     const double hour_angle_rad = d2r(topo_local_hour_angle);
-    const double eps_0 = r2d(asin(sin(lat_rad)*sin(sun_decl_rad) +
-                                  cos(lat_rad)*cos(sun_decl_rad)*cos(hour_angle_rad)));
+    const double eps_0 = r2d(asin(sin(lat_rad)*sin(sun_decl_rad) + cos(lat_rad)*cos(sun_decl_rad)*cos(hour_angle_rad)));
+
     return eps_0;
 }
 
-auto topocentric_zenith_angle(const double eps_0,
-                              const double P,
-                              const double T) {
+inline std::pair<double, double> topocentric_zenith_angle(const double eps_0, const double P, const double T) {
     // Equation (42)
     const double deps = P/1010*283/(273 + T)*1.02/(60.0*tan(d2r(eps_0 + 10.3/(eps_0 + 5.11))));
     // Equation (43)
     const double topocentric_elevation_angle = eps_0 + deps;
     // Equation (44)
     const double topocentric_zenith_angle = 90 - topocentric_elevation_angle;
+
     return std::make_pair(topocentric_elevation_angle, topocentric_zenith_angle);
 }
 
-auto topocentric_astronomers_and_azimuth_angle(const double topo_local_hour_angle,
-                                               const double geographical_latitude,
-                                               const double topo_sun_declination) {
+inline std::pair<double, double> topocentric_astronomers_and_azimuth_angle(
+    const double topo_local_hour_angle, const double geographical_latitude, const double topo_sun_declination
+) {
     // Equation (45)
     const double h_rad = d2r(topo_local_hour_angle);
     const double lat_rad = d2r(geographical_latitude);
     const double sun_decl_rad = d2r(topo_sun_declination);
-    const double Gamma = limit_degrees(r2d(atan2(sin(h_rad),
-                                    cos(h_rad)*sin(lat_rad) - tan(sun_decl_rad)*cos(lat_rad))));
+    const double Gamma = limit_degrees(
+        r2d(atan2(sin(h_rad), cos(h_rad)*sin(lat_rad) - tan(sun_decl_rad)*cos(lat_rad)))
+    );
     // Equation (46)
     const double Phi = limit_degrees(Gamma + 180);
-    return std::make_tuple(Gamma, Phi);
+
+    return std::make_pair(Gamma, Phi);
 }
 
-double indicence_angle(const double topo_zenith_angle, const double surface_slope, const double surface_azimuth, const double astronomers_azimuth_angle) {
+inline double indicence_angle(
+    const double topo_zenith_angle,
+    const double surface_slope,
+    const double surface_azimuth,
+    const double astronomers_azimuth_angle
+) {
     // Equation (47)
     const double zenit_rad = d2r(topo_zenith_angle);
     const double slope_rad = d2r(surface_slope);
     const double astro_azimuth_rad = d2r(astronomers_azimuth_angle);
     const double surf_azimuth_rad = d2r(surface_azimuth);
-    const double I = r2d(acos(cos(zenit_rad)*cos(slope_rad)
-                            + sin(slope_rad)*sin(zenit_rad)*cos(astro_azimuth_rad - surf_azimuth_rad)));
+    const double I = r2d(
+        acos(cos(zenit_rad)*cos(slope_rad) + sin(slope_rad)*sin(zenit_rad)*cos(astro_azimuth_rad - surf_azimuth_rad))
+    );
+
     return I;
 }
 
 template<typename collector_t>
-auto solar_position(const double julian_day,
-                    const double DT,
-                    const double geographic_latitude,
-                    const double geographic_longitude,
-                    const double masl,
-                    collector_t collector) {
+auto solar_position(
+    const double julian_day,
+    const auto DT,
+    const double geographic_latitude,
+    const double geographic_longitude,
+    const double masl,
+    collector_t collector
+) {
     // Note that the args should be in UT, and that |UT - UTC| < 1.0
 
     //const auto DT = Delta_T(year);
-    const auto julian_ephemeris_day = jde(julian_day, DT);
-    const auto julian_ephemeris_centry = jce(julian_ephemeris_day);
-    const auto julian_ephemeris_millennium = jme(julian_ephemeris_centry);
-    const auto L = heliocentric_longitude(julian_ephemeris_millennium);
-    const auto B = heliocentric_latitude(julian_ephemeris_millennium);
-    const auto R = heliocentric_radius_vector(julian_ephemeris_millennium);
-    const auto beta = geocentric_latitude(B);
-    const auto Theta = geocentric_longitude(L);
+    const double julian_ephemeris_day = jde(julian_day, DT);
+    const double julian_ephemeris_centry = jce(julian_ephemeris_day);
+    const double julian_ephemeris_millennium = jme(julian_ephemeris_centry);
+    const double L = heliocentric_longitude(julian_ephemeris_millennium);
+    const double B = heliocentric_latitude(julian_ephemeris_millennium);
+    const double R = heliocentric_radius_vector(julian_ephemeris_millennium);
+    const double beta = geocentric_latitude(B);
+    const double Theta = geocentric_longitude(L);
 
 
     const auto [longitude_nutation, obliquity_nutation] = nutation(julian_ephemeris_centry);
-    const auto epsilon = true_ecliptic_obliquity(julian_ephemeris_millennium, obliquity_nutation);
-    const auto delta_tau = aberration_correction(R);
-    const auto lambda = apparent_sun_longitude(Theta, longitude_nutation, delta_tau);
-    const auto nu = apparent_Greenwich_sidereal_time(julian_day,
-                                                     julian_ephemeris_centry,
-                                                     longitude_nutation,
-                                                     epsilon);
-    const auto alpha = geocentric_sun_right_ascension(lambda, epsilon, beta);
-    const auto delta = geocentric_sun_declination(lambda, epsilon, beta);
-    const auto H = observer_local_hour_angle(nu, alpha, geographic_longitude);
-    const auto [alpha_mark, delta_mark, H_mark]  = topocentric_values(R,
-                                                                      geographic_latitude,
-                                                                      masl,
-                                                                      alpha,
-                                                                      delta,
-                                                                      H);
-    const double e0 = uncorrected_topocentric_elevation_angle(geographic_latitude,
-                                                              delta_mark,
-                                                              H_mark);
-    const auto [Gamma, Phi] = topocentric_astronomers_and_azimuth_angle(H_mark,
-                                                                        geographic_latitude,
-                                                                        delta_mark);
+    const double epsilon = true_ecliptic_obliquity(julian_ephemeris_millennium, obliquity_nutation);
+    const double delta_tau = aberration_correction(R);
+    const double lambda = apparent_sun_longitude(Theta, longitude_nutation, delta_tau);
+    const double nu = apparent_Greenwich_sidereal_time(julian_day, julian_ephemeris_centry, longitude_nutation, epsilon);
+    const double alpha = geocentric_sun_right_ascension(lambda, epsilon, beta);
+    const double delta = geocentric_sun_declination(lambda, epsilon, beta);
+    const double H = observer_local_hour_angle(nu, alpha, geographic_longitude);
+    const auto [alpha_mark, delta_mark, H_mark] = topocentric_values(R, geographic_latitude, masl, alpha, delta, H);
+    const double e0 = uncorrected_topocentric_elevation_angle(geographic_latitude, delta_mark, H_mark);
+    const auto [Gamma, Phi] = topocentric_astronomers_and_azimuth_angle(H_mark, geographic_latitude, delta_mark);
+
     return collector(e0, Gamma, Phi, alpha_mark, delta_mark, H_mark);
 }
 
-auto corrected_solar_elevation(const double e0, const double P, const double T) {
+inline std::pair<double, double> corrected_solar_elevation(const double e0, const double P, const double T) {
     return topocentric_zenith_angle(e0, P, T);
 }
 
 template<typename collector_t, typename dt_calc_t>
-auto calendar_solar_position(unsigned int year,
-                             unsigned int month,
-                             double day,
-                             const double geographic_latitude,
-                             const double geographic_longitude,
-                             const double masl,
-                             collector_t collector,
-                             dt_calc_t cal_calc) {
-    const double DT = cal_calc(year, month, day);
+auto calendar_solar_position(
+    unsigned int year,
+    unsigned int month,
+    double day,
+    const double geographic_latitude,
+    const double geographic_longitude,
+    const double masl,
+    collector_t collector,
+    dt_calc_t cal_calc
+) {
+    const auto DT = cal_calc(year, month, day);
     const double julian_day = jd_from_cal(year, month, day);
-    return solar_position(julian_day,
-                          DT,
-                          geographic_latitude,
-                          geographic_longitude,
-                          masl, collector);
+    return solar_position(julian_day, DT, geographic_latitude, geographic_longitude, masl, collector);
 }
 
 template<typename collector_t, typename dt_calc_t>
-auto time_point_solar_position(const std::chrono::system_clock::time_point time_point,
-                                   const double geographic_latitude,
-                                   const double geographic_longitude,
-                                   const double masl,
-                                   collector_t collector,
-                                   dt_calc_t time_point_calc) {
-
+auto time_point_solar_position(
+    const std::chrono::system_clock::time_point time_point,
+    const double geographic_latitude,
+    const double geographic_longitude,
+    const double masl,
+    collector_t collector,
+    dt_calc_t time_point_calc
+) {
     const auto DT = time_point_calc(time_point);
-    const auto julian_day = jd_from_clock(time_point);
-    return solar_position(julian_day,
-                          DT,
-                          geographic_latitude,
-                          geographic_longitude,
-                          masl, collector);
+    const double julian_day = jd_from_clock(time_point);
+
+    return solar_position(julian_day, DT, geographic_latitude, geographic_longitude, masl, collector);
 }
 
 }
